python_module/requirements.txt
# Python dependencies for BT Bridge module

pyserial>=3.5
bleak>=0.21.0
